decimal is float [closes #1810]

+++ packages/TypeDeclaration/src/PropertyTypeInferer/DoctrineColumnPropertyTypeInferer.php
@@ -28,7 +28,7 @@ final class DoctrineColumnPropertyTypeInferer implements PropertyTypeInfererInte
         'int' => 'int',
         'integer' => 'int',
         'bigint' => 'int',
-        'decimal' => 'int',
+        'decimal' => 'float',
         'numeric' => 'int',
         'float' => 'float',
         'double' => 'float',
